requirement.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
peewee==3.17.0